requirements.txt
asn1crypto==0.24.0
attrs==18.1.0
Automat==0.6.0
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2018.4.16
cffi==1.11.5
chardet==3.0.4
click==6.7
constantly==15.1.0
cryptography==2.2.2
cssselect==1.0.3
geoip2==2.9.0
gitdb2==2.0.3
GitPython==2.1.10
hyperlink==18.0.0
idna==2.6
incremental==17.5.0
itsdangerous==0.24
lxml==4.2.1
MarkupSafe==1.0
maxminddb==1.4.0
parsel==1.4.0
pyasn1==0.4.3
pyasn1-modules==0.2.1
pycparser==2.18
PyDispatcher==2.0.5
pyOpenSSL==18.0.0
PySocks==1.6.8
queuelib==1.5.0
requests==2.18.4
service-identity==17.0.0
six==1.11.0
smmap2==2.0.3
Twisted==18.4.0
urllib3==1.22
w3lib==1.19.0
Werkzeug==0.14.1
zope.interface==4.5.0
